# Replace vars with consts in example template