… internal/proto/examples/scalars_canoto.go