Not sure why the other benchmarks would be impacted... probably chrome running in the background or something during the second run... But using constants significantly improves the performance. 🚀 
```
goos: darwin
goarch: arm64
pkg: github.com/StephenButtolph/canoto/internal/proto/examples
                                        │   old.txt   │               new.txt               │
                                        │   sec/op    │   sec/op     vs base                │
Scalars_MarshalCanoto-12                  92.08n ± 2%   70.05n ± 1%  -23.92% (p=0.000 n=10)
Scalars_UnmarshalCanoto/unsafe=false-12   143.8n ± 2%   145.8n ± 1%   +1.39% (p=0.050 n=10)
Scalars_UnmarshalCanoto/unsafe=true-12    111.0n ± 2%   116.2n ± 1%   +4.73% (p=0.001 n=10)
Scalars_MarshalProto-12                   192.1n ± 0%   203.8n ± 2%   +6.12% (p=0.000 n=10)
Scalars_UnmarshalProto-12                 181.0n ± 4%   186.4n ± 4%   +2.96% (p=0.003 n=10)
geomean                                   138.6n        135.2n        -2.47%
```

Serialization now takes `35`% of the time that proto does.